ai_services/pyproject.toml
[project]
name = "ai-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.5.8",
    "fastapi>=0.115.12",
    "pydantic[email]>=2.11.5",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "supabase>=2.15.2",
    "uvicorn>=0.34.3",
    "sqlalchemy>=2.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
]
